changed unsigned long to uint32_t

+++ Core/Inc/charlieplex.h
@@ -5,7 +5,7 @@
 
 void Charlieplex_SetLED(GPIO_TypeDef **ports, int num_ports, uint16_t *pins, int num_pins, int x, int y, uint8_t state);
 void Charlieplex_Clear(GPIO_TypeDef **ports, int num_ports, uint16_t *pins, int num_pins);
-void Charlieplex_Display(GPIO_TypeDef **ports, int num_ports, uint16_t *pins, int num_pins, uint8_t matrix[num_pins - 1][num_pins], unsigned long pixel_delay_us);
+void Charlieplex_Display(GPIO_TypeDef **ports, int num_ports, uint16_t *pins, int num_pins, uint8_t matrix[num_pins - 1][num_pins], uint32_t pixel_delay_us);
 
-void Delay_us(unsigned long us);
+void Delay_us(uint32_t us);
 #endif
